Move XCUIElementQuery utils to dedicated file in UITestsFoundation
