+++ WordPress/UITestsFoundation/BaseScreen.swift
@@ -112,21 +112,6 @@ extension BaseScreen {
     }
 }
 
-// TODO: This might be better suited in the XCUITestHelpers frameworks.
-public extension XCUIElementQuery {
-
-    var allElementsShareCommonXAxis: Bool {
-        let elementXPositions = allElementsBoundByIndex.map { $0.frame.minX }
-
-        // Use a set to remove duplicates – if all elements are the same, only one should remain
-        return Set(elementXPositions).count == 1
-    }
-
-    var lastMatch: XCUIElement? {
-        return self.allElementsBoundByIndex.last
-    }
-}
-
 // TODO: This should go into a UIDevice extension (eg: `UIDevice.current.isPad`)
 public var isIpad: Bool {
     return UIDevice.current.userInterfaceIdiom == .pad
